requirements.txt
matplotlib==3.4.3
numpy==1.21.2
pandas==1.3.2
Pillow==8.3.2
torch==1.9.0
torchvision==0.10.0
